fix(friendProgram): cleanup and add comments on methods

+++ store/friends/actions.ts
@@ -348,14 +348,18 @@ export default {
       }
     })
 
-    friendsProgram.addEventListener(FriendsEvents.REQUEST_DENIED, (account) => {
-      if (account) {
-        commit(
-          'removeOutgoingRequest',
-          friendAccountToOutgoingRequest(account, null).requestId,
-        )
-      }
-    })
+    friendsProgram.addEventListener(
+      FriendsEvents.REQUEST_DENIED,
+      async (account) => {
+        if (account) {
+          const userInfo = await usersProgram.getUserInfo(account.from)
+          dispatch(
+            'removeFriendRequest',
+            friendAccountToOutgoingRequest(account, userInfo),
+          )
+        }
+      },
+    )
 
     friendsProgram.addEventListener(
       FriendsEvents.REQUEST_REMOVED,
@@ -380,7 +384,6 @@ export default {
     { commit }: ActionsArguments<FriendsState>,
     { friendToKey }: CreateFriendRequestArguments,
   ) {
-    console.log('friendrequest')
     const $SolanaManager: SolanaManager = Vue.prototype.$SolanaManager
     const $Crypto: Crypto = Vue.prototype.$Crypto
     const $TextileManager: TextileManager = Vue.prototype.$TextileManager
@@ -394,11 +397,13 @@ export default {
 
     const friendsProgram: FriendsProgram = new FriendsProgram($SolanaManager)
 
-    const friendRequestKey = await friendsProgram.requestKey(friendToKey)
-
-    let friendAccountInfo = await friendsProgram.getFriendAccount(
-      friendRequestKey,
+    const { publicKey: friendFromKey } = friendsProgram.getPayer()
+    const accountKeys = await friendsProgram.computeAccountKeys(
+      friendFromKey,
+      friendToKey,
     )
+
+    let friendAccountInfo = await friendsProgram.getAccount(accountKeys.request)
     if (friendAccountInfo) {
       if (friendAccountInfo.status === FriendStatus.PENDING) {
         throw new Error(FriendsError.REQUEST_ALREADY_SENT)
@@ -417,16 +422,22 @@ export default {
       textilePublicKey,
     )
 
-    await friendsProgram.makeRequest(friendToKey, encryptedTextilePublicKey)
+    const tx = await friendsProgram.makeRequest(
+      accountKeys.request,
+      accountKeys.first,
+      accountKeys.second,
+      encryptedTextilePublicKey,
+    )
 
-    friendAccountInfo = await friendsProgram.getFriendAccount(friendRequestKey)
-
-    if (friendAccountInfo) {
-      const userInfo = await usersProgram.getUserInfo(friendAccountInfo.to)
-      commit(
-        'addOutgoingRequest',
-        friendAccountToOutgoingRequest(friendAccountInfo, userInfo),
-      )
+    if (tx) {
+      friendAccountInfo = await friendsProgram.getAccount(accountKeys.request)
+      if (friendAccountInfo) {
+        const userInfo = await usersProgram.getUserInfo(friendAccountInfo.to)
+        commit(
+          'addOutgoingRequest',
+          friendAccountToOutgoingRequest(friendAccountInfo, userInfo),
+        )
+      }
     }
   },
 
@@ -456,11 +467,7 @@ export default {
     commit('updateIncomingRequest', { ...friendRequest, pending: true })
     const { account } = friendRequest
 
-    const friendAccountKey = await friendsProgram.requestKey(
-      new PublicKey(account.from),
-    )
-
-    const friendFromKey = friendRequest.account.from
+    const friendAccountKey = new PublicKey(friendRequest.requestId)
 
     const friendCheck = await friendsProgram.getParsedFriend(
       computedFriendAccountKey,
@@ -477,12 +484,14 @@ export default {
       textilePublicKey,
     )
 
-    await friendsProgram.acceptRequest(
+    const tx = await friendsProgram.acceptRequest(
       friendAccountKey,
-      encryptedIdentityPublicKey,
+      encryptedTextilePublicKey,
     )
 
-    dispatch('fetchFriendDetails', account)
+    if (tx) {
+      dispatch('fetchFriendDetails', account)
+    }
   },
   /**
    * @method denyFriendRequest DocsTODO
@@ -508,16 +517,16 @@ export default {
 
     const { account } = friendRequest
 
-    const friendAccountKey = await friendsProgram.requestKey(
-      new PublicKey(account.from),
-    )
+    const friendAccountKey = new PublicKey(friendRequest.requestId)
 
-    await friendsProgram.denyRequest(friendAccountKey)
+    const tx = await friendsProgram.denyRequest(friendAccountKey)
 
-    commit(
-      'removeIncomingRequest',
-      friendAccountToIncomingRequest(account, null).requestId,
-    )
+    if (tx) {
+      commit(
+        'removeIncomingRequest',
+        friendAccountToIncomingRequest(account, null).requestId,
+      )
+    }
   },
 
   /**
@@ -545,16 +554,14 @@ export default {
     const { account } = friendRequest
     const friendAccountKey = new PublicKey(friendRequest.requestId)
 
-    const friendAccountKey = await friendsProgram.requestKey(
-      new PublicKey(account.to),
-    )
+    const tx = await friendsProgram.removeRequest(friendAccountKey)
 
-    await friendsProgram.removeRequest(friendAccountKey)
-
-    commit(
-      'removeOutgoingRequest',
-      friendAccountToOutgoingRequest(account, null).requestId,
-    )
+    if (tx) {
+      commit(
+        'removeOutgoingRequest',
+        friendAccountToOutgoingRequest(account, null).requestId,
+      )
+    }
   },
 
   /**
@@ -579,15 +586,17 @@ export default {
 
     const { account, address } = friend
 
-    const transactionId = await friendsProgram.removeFriend(
+    const tx = await friendsProgram.removeFriend(
       new PublicKey(account.accountId),
     )
 
-    commit('removeFriend', address)
-    if (this.app.router.currentRoute?.params?.address === address) {
-      this.app.router.replace('/chat/direct')
+    if (tx) {
+      commit('removeFriend', address)
+      if (this.app.router.currentRoute?.params?.address === address) {
+        this.app.router.replace('/chat/direct')
+      }
+      await db.friends.where('address').equals(address).delete()
     }
-    await db.friends.where('address').equals(address).delete()
   },
 
   /**
